Log stdout/stderr during error
Paunch does not currently show stdout/stderr from the docker cmd when an
error happens. Without this output, there is no way to tell what
actually caused the container command to fail.

Log the output with LOG.error if we get an unexpected return code so
that the actual error is displayed to the user.

Change-Id: I951cdf0cb2bfd87a22a239ff7522ab5bba9d6eb6
Closes-Bug: #1707948