Add log_tag option to container definition
This adds the --log-opt tag=<value> to the docker run command, which can
be quite useful depending on your logging needs.

In the TripleO case, we use journald as the default logging driver, and
this will get reflected in the logs by adding the tag to the
CONTAINER_TAG key in the log metadata.

bp logging-stdout-rsyslog

Change-Id: I3bfb1a5ce2d6a6f95519f5272348eae18d649712