Adjust tenacity mock
Recent changes in tenacity have broken how we mock the wait call for the test.
This change mocks out the inspect.getargspec call that is used in tenacity to
fix our unit tests for now. I think it might be beneficial to not actually test
the retry functionality in the paunch unit tests as it seems to be more
of a test for tenacity than paunch.

Change-Id: I37b09694de3ec9f31a7b945d19a83527c915b759
Related-Bug: #1788802