podman: create/delete systemd unit files when restart policy is used
This patch will create a basic systemd unit file that start/stop/status
any container that we manage with restart policy in paunch config.
It's only created when podman is the container runtime and when the
container is configured with a restart policy.

The systemd unit file will be removed when paunch removes the container.

KillMode=process is used so in the case of Neutron, we don't kill
children containers.

Note: if the policy is set to unless-stopped, we'll force the always
policy because unless-stopped doesn't exist in systemd.

Change-Id: I676e5fff3daecadba45efddff11f7afc602a50ef