runner: test if image exists before running inspect
When we run:
  $ podman inspect --type image --format exists <image>

Against a non-existing image, we end up with message like:
  error getting image <image>: unable to find <image> in local storage
  unable to find a name and tag match for <image> in repotags

The messages can confuse the operators, and we don't want that.

Podman will have a command called:
  $ podman image exists <image>

Which would return 0 if the image actually exists.
So before running the inspect command, we verify that the image
actually exists.
If the container CMD is Podman, the new function will test that the
image exists and if not the inspect won't run. No change for Docker as
the command isn't implemented in it.

Change-Id: Ie16ca7377e8e5e5c63d9630cf6d26e0bc8df6b1a
Related: https://github.com/containers/libpod/issues/1845