Prefix SystemD service by "tripleo_"
We now prefix the SystemD service so we can identify them better:
  e.g. systemctl list-unit-files | grep tripleo

It'll help to not conflict when rpms are installed on the host and
have the same service name as their container name.

For example haproxy rpm and haproxy container would have the same
service name so the prefix will help to not having this conflict
when removing the rpms during a cleanup by the operator.

Change-Id: I4bec5cfdb34c7422816c2d79ccaa5a499c783631