podman: only check image availability when inspecting image
When a paunch command (e.g. 'paunch list') requires inspection
of a container, it passes the container id to inspect, so we
cannot check for image availability at this time (we probably
don't need to anyway because the container exists).

Only check for image availability if the inspect function is
requested to inspect an image.

Change-Id: I8491eddaae172c8cf0b386d4731ab72dc113049c