Allow to pass some of systemd.exec flags
It may happen we have some shared directory mounted within
the container - if it is located in /run or in some other
tmpfs location, it won't persist across reboots.

SystemD provides the capability to create those directories, and
this patch allows to take advantage of this capacity.

Basically, we might add as many flag as we want.

Closes-Bug: #1814764
Change-Id: I5d2ef49ec205b7a43c488d4ceef0724e5ef2b6cc