Increase and randomize healthcheck interval to avoid cpu congestion.
All the healthcheck timers are triggered at the same time which causes
cpu congestion.  With the default interval of 30s this cause
unnecessary stress on the environment.

We increase the default interval and randomize the trigger to 3/4 of
the timer. Tasks, on average and by default, should thus still be
triggered every minute, but ranging from 15s to 1m45s.

Change-Id: Ib11b08b5d0dec03db9740520f05fe8f1ff962192
Closes-Bug: #1814870