systemd: switch KillMode to 'none'
When running the services with KillMode=process, there is a race
condition between ExecStop and the command specified in ExecStart.
The ExecStop seems faster and the container is killed then cleaned up.
However podman started by ExecStart is still running and systemd kills
it as soon as the ExecStop finished.

Since we rely on Podman to manage the containers & processes, let's
switch to KillMode=none.

Credits to Giuseppe Scrivano for explaining the root cause.

Change-Id: Icbf2b81477902e3d7ff9e064bf2408c2fc7e510e
Closes-Bug: #1821241