Use pidfile for services
Currently when we start the service we use podman start -a when we start
the service so systemd monitors that process for the service. The
problem is that this leaves podman start processes around which have a
bunch of threads and is attached to the STDOUT/STDIN.  Rather than
leaving this start process around, let's switch to a forking type in
system and specify a conmon pidfile for monitoring the container.  This
removes 500+ threads from the system.

Change-Id: Ie9225ce484b94b7fd59036168aaae58f25f1589d
Co-Authored-By: Emilien Macchi <emilien@redhat.com>