Add missing link between healthcheck unit and main service
Until now, if we stop the main service, we might encounter an issue
where the healthcheck is running, and we end up with two error:
- systemctl stop <container> returns "1"
- container is in Exited(137) state

[root@undercloud ~]# systemctl stop tripleo_nova_compute
Job for tripleo_nova_compute.service canceled.
[root@undercloud ~]# echo $?
1
podman ps -a | grep nova_compute
(...) 27 minutes ago  Exited (137) 3 minutes ago nova_compute

Using the "PartOf" within the healthecheck ensures it's stopped when the
main service is stopped.
Adding the "Requires" on the main service unit allows to ensure the
healthcheck is started as a dependency.

Change-Id: I681fa0eb225d43ecbaa1b765f6761dd07031bce1
Depends-On: https://review.opendev.org/660060