Introduce the systemctl module to provide simple systemctl commands wrap
Centralize all the subprocess dedicated to call systemctl in a dedicated
wrapper to ensure straightforward usage and dedicated unit test less
in a functional approach and also reduce the technical debt.

This module can continue to evolve to reflect the real systemctl command
signatures and dedicated parameters and related checks if needed.

Change-Id: Ia0e278bf5731aa5d09739d54f4db64dbaf85d037