runner: refactor image_exist
iamge_exist only makes sense in podman where the inspect command
requires the image to be here.

Also, rather than returning a return code, return a bool to make the use
of image_exist simpler in the code.

This patch is preparing the code to have container_exist and re-use the
same logic later.

Change-Id: I9b7f1df0f466f8e5cd44f0c8e4465618fe4fc8a1