Optimize container CLI for getting unique names
Do not call it twice in the while{} loop. It is pointless to
double-check for a container existence after we have determined weither
its name is in use or not.

This decreases the pressure on the container system.

Change-Id: Ife65f082f389ad4a730b3f9cf07d32ac4be32b45
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>
(cherry picked from commit a6df749249b36d2c4aa1621892e778eb9d8248ab)