Fix mismatching fixed vs unique container names
Sometimes, like when redeploying in-place, containers for the "fixed"
service name might already existed, thus get the prefixed names. That
might create mismatches. For example the pidfile names may diverge by
the "fixzed" container service name vs its predictable prefixed unique
name.

Fix that by using the predictable unique names instead of the service
container names for the builder and paunch actions run,
debug/print-cmd that rely on it. This is achieved via a new parameter
for the real container name (a delegate) used for the "fixed" service
container name.

For podman builder, we use that delegate for conmon pidfile and logging
setup. So that now it always matches the PIDFile specified in the
systemd unit generated for that container.

For docker builder, we have no special uses for delegates, but we
support that parameter to simplify the code around (so that there will
be no need to wrap things with 'if cli == podman else...').

Closes-bug: #1839929

Change-Id: I5617e11f5d315f408d818e1ce47aa68f4a0d777a
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>