Fix idempotency on RHEL8
On RHEL8, podman is shipped by default with podman-docker which installs
an alias so the operators can still use "docker" CLI, but the command
will be converted to run "podman" instead:

  Emulate Docker CLI using podman. Create /etc/containers/nodocker to quiet msg

We don't want to remove the rpm or create /etc/containers/nodocker so
the operators have the warning and now we are moving to Podman by
default.

So we need to change Paunch to detect if Docker was actually running
before removing docker-managed containers, instead of using 'which
docker', we'll check for the actual docker socket if it exists.

Co-Authored-By: Cédric Jeanneret <cjeanner@redhat.com>
Change-Id: I628c2fe7e962412f0d7e1c3e044d02c5a2fabd77
(cherry picked from commit 2ed3e83fdefb965ca62234a1657fc95dc6a27e99)