Add missing ExecReload in container service unit file
It may happen that we want to just reload the container. Before this
patch, it was a "stop and start", while podman has the "podman kill"
available, accepting the HUP signal.

Doing so allows other automated tools to actually just "reload" the
container as we would do for a standard service.

Change-Id: I5946a36e56dd1f3b49104deaef4ab69bada8e264