Do not always needlessly pull the images via podman
Via https://review.opendev.org/#/c/665731/ there was a tentative
reduction of podman inspect calls. The problem was that a podman
pull was performed even when the image already existed due to
a wrong if check, since image_exist() returns true/false and not
the return code of the podman image shell command.

Closes-Bug: #1855128
Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>

Change-Id: Ia0c71de6a0a1fe243a39dbd7b1458610b81bd6be