Exit 1 if a container fails and return the error
When a container's volumes failed validation, paunch still exited 0.
This caused the deployment to continue running even though not all
containers had been started.

This patch changes the rc to 1 when a container's volumes fail
validation and the container can't be started. The error message is also
returned in stderr so that it's available to the paunch ansible module
and will be seen in the deployment output.

Depends-On: I1f062b8b9f936e6fbf2febf64244e91b59b8ba1b
Change-Id: I67860a79572c0ff4dcaca9ec9597c41f56792fca
Closes-Bug: #1855444
(cherry picked from commit 3ab0936c03dfdd8b12fb931d3a3480982a76574e)