builder: include environment when running an exec
Some container execs have an environment, let's make sure they are
included when running exec.

NOTE: This is for Rocky and below as docker was the only supported
method in these versions. Additionally in Rocky and below, the
environment was a list and not a dictionary so we only support a list
version for the environment.

Closes-Bug: #1855932
Change-Id: Ic2e2c2d50f5883f7db28768ba215e74bcbf9fd8b
(cherry picked from commit b5be45067ac1e7d3b08458e2943e8de046aeee7b)