Update zuul layout to reflect recent changes in TripleO
Paunch is now disabled by default except on a few jobs when Pacemaker is
still running with Docker; which is the case of scenario004.

Let's remove the jobs where Paunch isn't exercised, to save resources.
And also add the 'irrelevant-files' parameter to skip the job when not
needed.

Note: it's not defined in that file but the OVB jobs running on CentOS7
also need to be executed until they are moved to CentOS8; since
Pacemaker and Docker are used; therefore Paunch is used to manage the
non-HA containers.

Change-Id: Ie8d2515b46f70b5ca39625bd0272d933f3b0ffec