Do not set cpuset-cpus if cconfig['cpuset_cpus'] == 'all'
In the case of nova_libvirt container, we want to use all CPUs that are
reported online.
Rather than computing the list with Python (which has proven to be
problematic on PPC), let the container engine figuring it out by itself
like it was the case before.

Change-Id: I5d1b88c90dbc4114d996008a407cd1dd9a6eb9da
Closes-Bug: #1868135
(cherry picked from commit 9b6276512141693bddd064c915b08c1d027e5985)