Make sure failed containers get stopped by systemd
When the main PID (i.e., common) of a container is killed because of
some reasons, systemd won't execute ExecStop command.
Current podman doesn't detect this failure ang recognize that container
is still running and this causes failure when systemd tries to restart
the container.

This patch introduces ExecStopPost configuration into systemd unit
files so that stop operation is executed even when a container fails
because of killed main process. The stale container should be cleaned
up by ExecStopPost task before systemd tries to restart it.

Note that the similar change has been introcuded to "podman generate
systemd" command already[1].
[1] https://github.com/containers/libpod/commit/e5c3432944245a740ed443803c654dcc9c3757f0

Change-Id: Ie2cb3bb63138b70019f86b88066697e98e333669
Closes-Bug: #1884866