[ROCKY-Only] Trigger container update on image id update
NOTE: Rocky is missing a few backports so this patch is rocky specific.

If you reuse a container tag for a new image, paunch does not currently
rebuild the container to get the new image.  This change adds logic to
the updated check to also check the image id defined in the container
instance against the image id for the image to see if the container
image has changed.

Closes-Bug: #1895974
Change-Id: I2710497282eb4d6fa8f68dc6d50044c1ce68eb35
(cherry picked from commit 4ae7d886db604f1ac00102803a5ea1fa2b3a479b)