[QUEENS Only] Always do a image pull
In queens we use paunch to manage the images and containers. Container
images are not updated if the image has changed but the container tag
has not because we simply check if the container exists on disk. This
change drops the exists check and just issues a pull for all containers
to ensure they are up to date. In newer versions we always pre-fetch the
containers as part of the deployment so we don't need to do this
image management in paunch (or ansible).

Change-Id: Icb5b949c0d9cd41328e1af3d4ec3066d6971c6fd
Closes-Bug: #1895974