Actually fail if a stop or remove call fails
As part of the container rename actions, attempt to stop and remove
containers without actually checking if they exist. Additionally we
don't actually fail if these command fail which leads to other errors
later that may be unrelated to the core issue. This change adds
additional checks around the stop and remove container actions to ensure
that the container is running before stopping and exists before
removing.

Change-Id: I82172f5a2d4683509c802c0f14d571a40e21e6cc
Closes-Bug: #1907833