Ensure systemd services are activated
When deployment times out during running the paunch command, containers
might be left in an incomplete status. Especially when podman is used
and containers are managed by systemd, sometimes paunch succeeds to
create a container but it is aborted before creating and enabling
the systemd service for that container.

This change makes paunch check status of systemd services, to ensure
any disabled/stopped/missing services are properly recovered by next
run.

Closes-Bug: #1955056
Change-Id: I1713d55318834cd93a0a045970aa4800aed4cd68