Ensure systemd services are activated
When deployment times out during running the paunch command, containers
might be left in an incomplete status. Especially when podman is used
and containers are managed by systemd, sometimes paunch succeeds to
create a container but it is aborted before creating and enabling
the systemd service for that container.

This change makes paunch check status of systemd services, to ensure
any disabled/stopped/missing services and remove containers with
incomplete systemd services so that the containers and the associated
systemd services are recreated and reconfigured.

Closes-Bug: #1955056
Change-Id: I7a1b5514cb4b80999d2fc884d19b6e62f94ba3c5