Ensure the dependent service is started in advance
Currently when dependent service is defined using the depends_on key,
paunch adds 'Wants' so that the dependent service is also started.

However this attribute does not guarantee that the dependent service is
started BEFORE the target service is started.

Because current usage of depends_on expect the dependent service is
started before the target service, this change adds 'After' so that
the ordering is properly enforced.

Note:
This change is pseudo-backport of tripleo-ansible patch, to implement
the same change in stable/train which uses paunch by default to start
up containers.

Also, pins virtualenv to avoid current py37 failure and drops
releasenote job.

Closes-Bug: #1970354
Change-Id: I2fd8f08c1dc96f37ae7588b68f57f5417d21c7d8